+++ Source/ShooTingGame/FuselageStatus.cpp
@@ -3,7 +3,7 @@
 
 #include "FuselageStatus.h"
 
-FuselageStatus::FuselageStatus(const int32 MaxHP, const float Speed, const int32 AttackPower)
+FuselageStatus::FuselageStatus(const float MaxHP, const float Speed, const int32 AttackPower)
 	: m_max_hp(MaxHP), m_speed(Speed), m_attack_power(AttackPower) {}
 
 FuselageStatus::~FuselageStatus()
